DOT.js: listen engine fixed, test payments add; Prestashop: bugs fixed

+++ make_zip.php
@@ -2,22 +2,23 @@
 
 <?php
 
-$g=glob("./*.txt");
+$g=glob("./*_release");
 if(empty($g)) die("no files");
 
+do_store("./common");
 foreach($g as $file) do_store($file);
 
-function do_store($file) {
+function do_store($dir) { $file=$dir."/".str_replace('_release','',basename($dir)).".txt";
+    echo "==================== ".$file."\n";
+    if(!is_file($file)) die("File not found: ".$file);
+
     $s=file_get_contents($file);
-//    echo $s."\n";
-//    return;
 
     $TO="/tmp/UPUPO";
-    exec("rm -r \"".$TO."\"");
+    if(is_dir($TO)) exec("rm -r \"".$TO."\"");
 
-    $DIR="./ps_dotpayment";
-    $ZIP=$_SERVER['PWD']."/ps_dotpayment_".date("Y-m-d_H-i").".zip";
-
+    $DIR="/dev/null"; // ps_dotpayment";
+    // $ZIP=$_SERVER['PWD']."/ps_dotpayment_".date("Y-m-d_H-i").".zip";
 
     $s=explode("\n",$s); foreach($s as $l) {
         if(!strstr($l,'|') || strstr($l,'#') || substr($l,0,1)=='@') continue;
@@ -37,16 +38,28 @@ function do_store($file) {
 	}
 
 	if(!strstr($from,'://')) $from=$DIR.'/'.$from;
-	$to=$TO.'/'.$to;
+
+	if(strpos($to,'common/temp-upload')===false) $to=$TO.'/'.$to;
+	else $to="./".$to;
+
         if(basename($to)=='*') $to=str_replace('*',basename($from),$to);
-        echo "copy: [$from] -> [$to]\n";
-        $dd=explode('/',dirname($to)); if($dd[0]=='') unset($dd[0]);
-        $d=''; foreach($dd as $l) { $d.='/'.$l; if(!is_dir($d)) { mkdir($d); chmod($d,0777); }  }
+        // echo "copy: [$from] -> [$to]\n";
+	echo " $to\n";
+        $dd=explode('/',dirname($to));
+
+        $d=''; foreach($dd as $l) {
+	    $d.=$l;
+	    if($d!='/' && $d!='' && !is_dir($d)) {
+		// echo "mkdir(".$d.")\n";
+		mkdir($d);
+		chmod($d,0777);
+	    }
+	    $d.="/";
+	}
 
         if(!strstr($from,'://') && !is_file($from)) die("Error: file not found [".$from."]");
 
-        $o=file_get_contents($from);
-        if(empty($o)) die("Error: empty file [".$from."]");
+        $o=file_get_contents($from); if(empty($o)) die("Error: empty file [".$from."]");
 
         foreach($s as $l) {
 	    if(substr($l,0,1)!='@') continue;
@@ -60,12 +73,15 @@ function do_store($file) {
 	chmod($to,0666);
     }
 
-    echo "create: [$ZIP]\n";
+    if(isset($ZIP)) {
+        echo "ZIP: $ZIP\n";
+        if(is_file($ZIP)) unlink($ZIP);
+        exec("cd \"".$TO."\"; zip -r \"".$ZIP."\" \"./\"");
+        chmod($ZIP,0666);
+        copy($ZIP,"/tmp/".basename($ZIP));
+    }
 
-    if(is_file($ZIP)) unlink($ZIP);
-    exec("cd \"".$TO."\"; zip -r \"".$ZIP."\" \"./\"");
-    chmod($ZIP,0666);
-    copy($ZIP,"/tmp/".basename($ZIP));
+    echo "OK\n";
 }
 
 ?>
